Changed setup.py to pyproject.toml, formatted with black
