+++ pyproject.toml
@@ -1,0 +1,25 @@
+[build-system]
+requires = ["setuptools", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name= "MonetioCL"
+version = "0.1.0"
+description = "A CLI tool for adapting Chilean Observation Networks data to Melodies-Monet usable data."
+readme = "README.md"
+license = {text = "MIT"}
+requires-python = ">=3.12"
+authors = [{ name = "Antonio Andrés Torga Mellado", email = "antonio.torga@ug.uchile.cl" }]
+dependencies = [
+        "typer>=0.12",
+        "dask",
+        "xarray",
+        "pyarrow",
+        "httpx",
+]
+
+[project.scripts]
+MonetioCL= "MonetioCL.cli:app"
+
+[tool.black]
+line-length = 88
