+++ src/MonetioCL/config/config.json
@@ -1,0 +1,82 @@
+{
+    "defaults": {
+        "raw": {
+            "path": "input",
+            "format": "csv"
+        },
+        "intermediate": {
+            "path": "intermediate_results",
+            "format": "csv"
+        },
+        "output": {
+            "path": "output",
+            "format": "csv"
+        },
+        "stations": {
+            "path": ""
+        }
+    },
+    "observations": {
+        "melodies-monet-format": {
+            "dimensions": ["time", "siteid", "latitude", "longitude"]
+        },
+        "sinca": {
+            "data": {
+                "url" : "",
+                "info": {
+                    "var_string":null,
+                    "st_code": null,
+                    "year": null,
+                    "month": null,
+                    "user": null,
+                    "password": null
+                }
+                    },
+            "stations": {
+                "url" : "https://climatologia.meteochile.gob.cl/application/servicios/getEstacionesRedEma?usuario={}&token={}",
+                "info": {
+                    "user": null,
+                    "password": null
+                }
+            },
+
+            "dimensions" : {
+                "latitude": "latitud",
+                "longitude": "longitud",
+                "siteid": "codigoNacional",
+                "time": "momento"
+            }
+        },
+        "dmc": {
+            "data": {
+                "url" : "https://climatologia.meteochile.gob.cl/application/mensual/{}/{}/{}/{}?usuario={}&token={}",
+                "info": {
+                    "var_string":null,
+                    "st_code": null,
+                    "year": null,
+                    "month": null,
+                    "user": null,
+                    "password": null
+                }
+            },
+            "stations": {
+                "url" : "https://climatologia.meteochile.gob.cl/application/servicios/getEstacionesRedEma?usuario={}&token={}",
+                "info": {
+                    "user": null,
+                    "password": null
+                }
+            },
+            "dimensions" : {
+                "latitude": "latitud",
+                "longitude": "longitud",
+                "siteid": "codigoNacional",
+                "time": "momento"
+            }
+        }
+    },
+    "models":{
+        "melodies-monet-format": {
+            "dimensions": ["time", "x", "y", "latitude", "longitude"]
+        }
+    }
+}
